Оправени роли ученик да не може да пипа

+++ Controllers/ClassesController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -19,41 +20,26 @@ namespace GradingSystem.Controllers
             _context = context;
         }
 
-        // GET: Classes
         public async Task<IActionResult> Index()
         {
             return View(await _context.Classes.ToListAsync());
         }
 
-        // GET: Classes/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var @class = await _context.Classes
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (@class == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var @class = await _context.Classes.FirstOrDefaultAsync(m => m.Id == id);
+            if (@class == null) return NotFound();
             return View(@class);
         }
 
-        // GET: Classes/Create
+        [Authorize(Roles = "Admin")]
         public IActionResult Create()
         {
             return View();
         }
 
-        // POST: Classes/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
+        [HttpPost, ValidateAntiForgeryToken, Authorize(Roles = "Admin")]
         public async Task<IActionResult> Create([Bind("Id,Name")] Class @class)
         {
             if (ModelState.IsValid)
@@ -65,34 +51,19 @@ namespace GradingSystem.Controllers
             return View(@class);
         }
 
-        // GET: Classes/Edit/5
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var @class = await _context.Classes.FindAsync(id);
-            if (@class == null)
-            {
-                return NotFound();
-            }
+            if (@class == null) return NotFound();
             return View(@class);
         }
 
-        // POST: Classes/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
+        [HttpPost, ValidateAntiForgeryToken, Authorize(Roles = "Admin")]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] Class @class)
         {
-            if (id != @class.Id)
-            {
-                return NotFound();
-            }
-
+            if (id != @class.Id) return NotFound();
             if (ModelState.IsValid)
             {
                 try
@@ -102,49 +73,28 @@ namespace GradingSystem.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!ClassExists(@class.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!ClassExists(@class.Id)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
             return View(@class);
         }
 
-        // GET: Classes/Delete/5
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var @class = await _context.Classes
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (@class == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var @class = await _context.Classes.FirstOrDefaultAsync(m => m.Id == id);
+            if (@class == null) return NotFound();
             return View(@class);
         }
 
-        // POST: Classes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
+        [HttpPost, ActionName("Delete"), ValidateAntiForgeryToken, Authorize(Roles = "Admin")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var @class = await _context.Classes.FindAsync(id);
-            if (@class != null)
-            {
-                _context.Classes.Remove(@class);
-            }
-
+            if (@class != null) _context.Classes.Remove(@class);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
